# docs/swagger.yaml
basePath: /
definitions:
  github_com_raozhaizhu_go-estate_internal_domain_user.Role:
    enum:
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-varnames:
    - RoleUser
    - RoleVip
    - RoleAdmin
...
  github_com_raozhaizhu_go-estate_internal_service_user.DTO:
    properties:
      avatar_key:
        example: default_avatar.png
        type: string
      email:
        example: Bob@test.com
        type: string
      id:
        example: 1
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/github_com_raozhaizhu_go-estate_internal_domain_user.Role'
        example: 1
      username:
        example: Bob
        type: string
    type: object
  github_com_raozhaizhu_go-estate_pkg_api.AuthErrorResult:
    properties:
      code:
        example: 40100
        type: integer
      msg:
        example: 未授权，请检查令牌
        type: string
    type: object
  github_com_raozhaizhu_go-estate_pkg_api.ClientErrorResult:
    properties:
      code:
        example: 40000
        type: integer
      msg:
        example: 参数校验失败
        type: string
    type: object
  github_com_raozhaizhu_go-estate_pkg_api.ConflictErrorResult:
    properties:
      code:
        example: 40900
        type: integer
      msg:
        example: 数据冲突，操作无法完成
        type: string
    type: object
  github_com_raozhaizhu_go-estate_pkg_api.NotFoundErrorResult:
    properties:
      code:
        example: 40400
        type: integer
      msg:
        example: 资源不存在
        type: string
    type: object
  github_com_raozhaizhu_go-estate_pkg_api.ServerErrorResult:
    properties:
      code:
        example: 50000
        type: integer
      msg:
        example: 服务器内部发生错误
        type: string
    type: object
...
  github_com_raozhaizhu_go-estate_pkg_api.SuccessResult-github_com_raozhaizhu_go-estate_internal_service_user_DTO:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/github_com_raozhaizhu_go-estate_internal_service_user.DTO'
      msg:
        example: 操作成功
        type: string
    type: object
...
  github_com_raozhaizhu_go-estate_pkg_api.SuccessResult-internal_controller_user_AvatarUploadData:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/internal_controller_user.AvatarUploadData'
      msg:
        example: 操作成功
        type: string
    type: object
...
  internal_controller_user.AvatarUploadData:
    properties:
      form_data:
        additionalProperties:
          type: string
        type: object
      object_key:
        type: string
      post_url:
        type: string
    type: object
  internal_controller_user.CreateUserRequest:
    properties:
      avatar_key:
        example: default_avatar.png
        type: string
      email:
        example: Bob@test.com
        type: string
      password:
        example: "12345678"
        maxLength: 16
        minLength: 8
        type: string
      username:
        example: Bob
        maxLength: 32
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  internal_controller_user.UpdateUserRequest:
    properties:
      avatar_key:
        example: default_avatar.png
        type: string
      email:
        example: Bob@test.com
        type: string
      password:
        example: "12345678"
        maxLength: 16
        minLength: 8
        type: string
      username:
        example: Bob
        maxLength: 32
        minLength: 3
        type: string
    required:
    - username
    type: object
host: localhost
info:
  contact: {}
  description: 房地产项目的后端 API 接口文档
  title: Go Estate API 文档
  version: "1.0"
paths:
...
  /api/v1/user:
    post:
      consumes:
      - application/json
      description: 公开访问接口，用于普通用户的注册。任何人均可尝试创建。
      parameters:
      - description: 注册参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controller_user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.SuccessResult-github_com_raozhaizhu_go-estate_internal_service_user_DTO'
        "400":
          description: '参数校验失败 (HTTP 返回 200, code: 400xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.ClientErrorResult'
        "409":
          description: '用户名或邮箱已存在 (HTTP 返回 200, code: 409xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.ConflictErrorResult'
        "500":
          description: '服务器内部错误 (HTTP 返回 500, code: 500xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.ServerErrorResult'
      security:
      - DeviceIDAuth: []
      - UserAgentAuth: []
      summary: 注册普通用户
      tags:
      - User (用户管理)
  /api/v1/user/{username}:
    get:
      description: 获取指定用户的详细资料。（安全策略：仅限用户本人访问自己的数据，或者由 Admin 管理员强行调取）
      parameters:
      - description: 目标用户名
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.SuccessResult-github_com_raozhaizhu_go-estate_internal_service_user_DTO'
        "401":
          description: '未登录或越权非法访问 (HTTP 返回 200, code: 401xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.AuthErrorResult'
        "404":
          description: '该用户不存在 (HTTP 返回 200, code: 404xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.NotFoundErrorResult'
        "500":
          description: '服务器内部错误 (HTTP 返回 500, code: 500xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.ServerErrorResult'
      security:
      - BearerAuth: []
      - DeviceIDAuth: []
      - UserAgentAuth: []
      summary: 查询指定用户信息
      tags:
      - User (用户管理)
    patch:
      consumes:
      - application/json
      description: 局部修改用户字段。（安全策略：仅限用户本人操作，或者由 Admin 管理员代为修改）
      parameters:
      - description: 目标用户名
        in: path
        name: username
        required: true
        type: string
      - description: 需要更新的字段参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controller_user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.SuccessResult-github_com_raozhaizhu_go-estate_internal_service_user_DTO'
        "400":
          description: '参数解析错误或没有任何可更新的字段 (HTTP 返回 200, code: 400xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.ClientErrorResult'
        "401":
          description: '未登录或越权非法操作 (HTTP 返回 200, code: 401xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.AuthErrorResult'
        "404":
          description: '该用户不存在 (HTTP 返回 200, code: 404xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.NotFoundErrorResult'
        "500":
          description: '服务器内部错误 (HTTP 返回 500, code: 500xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.ServerErrorResult'
      security:
      - BearerAuth: []
      - DeviceIDAuth: []
      - UserAgentAuth: []
      summary: 修改指定用户信息
      tags:
      - User (用户管理)
  /api/v1/user/avatar/upload-url:
    get:
      consumes:
      - application/json
      description: 获取用于直接上传头像到对象存储的预签名凭证及表单参数。（安全策略：仅限已登录用户操作）
      parameters:
      - description: 文件后缀名 (支持 png, jpg, jpeg, webp，默认 png)
        in: query
        name: ext
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.SuccessResult-internal_controller_user_AvatarUploadData'
        "400":
          description: '参数解析错误或格式不支持 (HTTP 返回 200, code: 400xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.ClientErrorResult'
        "401":
          description: '未登录或鉴权失败 (HTTP 返回 200, code: 401xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.AuthErrorResult'
        "500":
          description: '服务器内部错误 (HTTP 返回 500, code: 500xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.ServerErrorResult'
      security:
      - BearerAuth: []
      - DeviceIDAuth: []
      - UserAgentAuth: []
      summary: 获取头像预上传凭证
      tags:
      - User (用户管理)
  /api/v1/user/vip:
    post:
      consumes:
      - application/json
      description: 受保护接口，仅限 Admin 管理员权限调用，用于直接创建带有 VIP 身份的用户。
      parameters:
      - description: VIP 用户参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controller_user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.SuccessResult-github_com_raozhaizhu_go-estate_internal_service_user_DTO'
        "400":
          description: '参数校验失败 (HTTP 返回 200, code: 400xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.ClientErrorResult'
        "401":
          description: '未登录或角色权限不足 (HTTP 返回 200, code: 401xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.AuthErrorResult'
        "409":
          description: '用户名或邮箱已存在 (HTTP 返回 200, code: 409xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.ConflictErrorResult'
        "500":
          description: '服务器内部错误 (HTTP 返回 500, code: 500xx)'
          schema:
            $ref: '#/definitions/github_com_raozhaizhu_go-estate_pkg_api.ServerErrorResult'
      security:
      - BearerAuth: []
      - DeviceIDAuth: []
      - UserAgentAuth: []
      summary: 管理员创建 VIP 用户
      tags:
      - User (用户管理)
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  DeviceIDAuth:
    in: header
    name: X-Device-ID
    type: apiKey
  UserAgentAuth:
    in: header
    name: User-Agent
    type: apiKey
swagger: "2.0"